Create a helpers class for commonly used methods
