+++ main.go
@@ -6,7 +6,6 @@ import (
 	"io/ioutil"
 	"log"
 	"net/http"
-	"strconv"
 	"strings"
 
 	"github.com/m7shapan/njson"
@@ -24,8 +23,8 @@ func show(pokemon app.Pokemon) {
 	fmt.Println("Name:", pokemon.Name)
 	fmt.Println("National Pokédex number:", pokemon.ID)
 
-	fmt.Println("Height:", convertDecimetersToMeters(pokemon.Height))
-	fmt.Println("Weight:", convertHectogramsToKilograms(pokemon.Weight))
+	fmt.Println("Height:", app.ConvertDecimetersToMeters(pokemon.Height))
+	fmt.Println("Weight:", app.ConvertHectogramsToKilograms(pokemon.Weight))
 
 	fmt.Println("------- Types -------")
 	for _, pokemonTypes := range pokemon.Types {
@@ -37,29 +36,6 @@ func show(pokemon app.Pokemon) {
 	fmt.Println("----- Abilities -----")
 	for _, pokemonAbilities := range pokemon.Abilities {
 		printPokemonAbilities(pokemonAbilities)
-	}
-}
-
-func convertDecimetersToMeters(height int) string {
-	if height >= 10 {
-		var heightAsFloat = float64(height)
-		var adjustedHeight = heightAsFloat / float64(10)
-		return fmt.Sprintf("%g", adjustedHeight) + "m"
-	} else {
-		return "0." + strconv.Itoa(height) + "m"
-	}
-}
-
-func convertHectogramsToKilograms(weight int) string {
-	if weight >= 10 {
-		var weightAsFloat = float64(weight)
-		var adjustedWeight = weightAsFloat / float64(10)
-		return fmt.Sprintf("%g", adjustedWeight) + "kg"
-	} else {
-		var weightAsFloat = float64(weight)
-		var adjustedWeight = weightAsFloat * 2.2
-		var weightAsPounds = adjustedWeight / float64(10)
-		return fmt.Sprintf("%.1f", weightAsPounds) + "lbs"
 	}
 }
